+++ static_precompiler/tests/test_management.py
@@ -53,3 +53,13 @@ def test_compilestatic_command(verbosity, capsys, monkeypatch, tmpdir):
         )
     else:
         assert stdout == ""
+
+
+@pytest.mark.skip("Re-enable when pytest-django>3.1.2 is released")
+@pytest.mark.django_db
+def test_ignore_dependencies_option(django_assert_num_queries, monkeypatch, tmpdir):
+
+    monkeypatch.setattr("static_precompiler.settings.ROOT", tmpdir.strpath)
+
+    with django_assert_num_queries(0):
+        management.call_command("compilestatic", ignore_dependencies=True)