+++ docs/conf.py
@@ -55,7 +55,7 @@ author = u'Andrey Fedoseev'
 # built documents.
 #
 # The short X.Y version.
-version = '1.7.1'
+version = '1.8'
 # The full version, including alpha/beta/rc tags.
 release = version